Added team agent. Simplified attacking environments. Open questions about assertions and design. Added attack to action space
