+++ examples/fighting_for_resources.py
@@ -2,13 +2,14 @@
 from matplotlib import pyplot as plt
 import numpy as np
 
-from admiral.component_envs.world import GridWorldEnv
+from admiral.component_envs.world import GridWorldEnv, WorldAgent
 from admiral.component_envs.movement import GridMovementEnv, GridMovementAgent
 from admiral.component_envs.resources import GridResourceEnv, GridResourceAgent
-from admiral.component_envs.attacking import GridAttackingEnv, AttackingTeamAgent
+from admiral.component_envs.attacking import GridAttackingTeamEnv, GridAttackingAgent
 from admiral.component_envs.death_life import DyingAgent, DyingEnv
+from admiral.component_envs.team import TeamAgent
 
-class FightForResourcesAgent(DyingAgent, AttackingTeamAgent, GridMovementAgent, GridResourceAgent):
+class FightForResourcesAgent(WorldAgent, DyingAgent, GridAttackingAgent, GridMovementAgent, GridResourceAgent, TeamAgent):
     pass
 
 class FightForResourcesEnv:
@@ -17,7 +18,7 @@ class FightForResourcesEnv:
         self.world = GridWorldEnv(**kwargs)
         self.resource = GridResourceEnv(**kwargs)
         self.movement = GridMovementEnv(**kwargs)
-        self.attacking = GridAttackingEnv(**kwargs)
+        self.attacking = GridAttackingTeamEnv(**kwargs)
         self.dying = DyingEnv(**kwargs)
 
         self.attacking_record = []
@@ -82,7 +83,7 @@ for _ in range(20):
         action_dict[agent_id] = {
             'move': agent.action_space['move'].sample(),
             'harvest': agent.action_space['harvest'].sample(),
-            'attack': np.random.randint(0, 2)
+            'attack': agent.action_space['attack'].sample(),
         }
     env.step(action_dict)
     env.render(fig=fig)
